refactor(core): split SpliceGraph core and parser modules
