+++ tests/test_integration_simple.py
@@ -1,8 +1,9 @@
 from pathlib import Path
 
+from SpliceGrapher.core.splice_graph import SpliceGraph
 from SpliceGrapher.formats.annotation_io import load_gene_models
+from SpliceGrapher.formats.parsers.splice_graph import SpliceGraphParser
 from SpliceGrapher.formats.writers.splice_graph import write_splice_graph_gff
-from SpliceGrapher.SpliceGraph import SpliceGraph, SpliceGraphParser
 from tests.helpers.idiffir_fixture_builder import build_fixture
