+++ tests/test_splicing_events.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 from SpliceGrapher.core.enums import AlternativeSplicingEvent
+from SpliceGrapher.core.splice_graph import SpliceGraph
 from SpliceGrapher.core.splicing_events import annotate_graph_events
-from SpliceGrapher.SpliceGraph import SpliceGraph
 
 
 def test_annotate_graph_events_marks_skipped_exon_for_bypass_motif() -> None:
